fix(portability): std::tanhf -> std::tanh (non-standard on GCC/libstdc++)
std::tanhf is an MSVC/libc++ extension; libstdc++ only provides the overloaded std::tanh.
Use std::tanh (its float overload handles the float arg). Last Linux compile blocker —
Windows + macOS already build + pass tests.

+++ mu-core/Audio/MultiModeFilter.cpp
@@ -91,7 +91,7 @@ void MultiModeFilter::process(juce::AudioBuffer<float>& buffer,
             {
                 const float y = chData[ch][i] * preGain;
                 // Positive: tanh.  Negative: x/(1−x) (slightly harder → even harmonics).
-                chData[ch][i] = (y >= 0.0f ? std::tanhf(y) : y / (1.0f - y)) * invGain;
+                chData[ch][i] = (y >= 0.0f ? std::tanh(y) : y / (1.0f - y)) * invGain;
             }
         }
     }
